requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
pillow
gradio